// src/api/models/AttachmentMetadataDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NOTE: This file is auto-generated by qswagger (https://github.com/Patrolin/qswagger)
 * Do NOT edit it manually
 */
export type AttachmentMetadataDto = {
    clientId: string;
    contentType: string;
    filename: string;
    id: string;
    messageId: string;
    size: number;
};