// src/api/models/OutboundChatMessage.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NOTE: This file is auto-generated by qswagger (https://github.com/Patrolin/qswagger)
 * Do NOT edit it manually
 */
import {AttachmentMetadataDto} from './AttachmentMetadataDto';

export type OutboundChatMessage = {
    attachments: AttachmentMetadataDto[];
    author: string;
    clientId: string;
    content: string;
    destination: string;
    destinationType: 'CHANNEL' | 'USER' | 'SELF';
    id: string;
    lastUpdateTimestamp: string;
    timestamp: string;
};